Initial Commit
Tests: to get a junit configuration correctly I had to declare a 'build'
in Gradle in addition to 'testCompile'.
I also had to declare 'org.eclipe.jetty.servlet' dependency because it
was breaking the compile, but I do not know who depends on it.
The standard output of the gradle test task is inefficient and
inadequate.  I have modified it for a better understanding of what is
happening.

Ensign + atom editor is a great setting to work with scala.

PrometheusSparkMetrics: I get the base code in this forum post
http://bit.ly/monitoring-spark-jobs.
I have modified it to save metric in seconds and change metrics name to
send the measure because it's a good pratice in prometheus community.